+++ code/drasil-data/drasil-data.cabal
@@ -12,7 +12,6 @@ library
     Data.Drasil.Constraints
     Data.Drasil.SentenceStructures
     Data.Drasil.SI_Units
-    Data.Drasil.Utils
     Data.Drasil.Concepts.Computation
     Data.Drasil.Concepts.Documentation
     Data.Drasil.Concepts.Thermodynamics
